Backend added to user tournaments
